+++ bindings/python/pyiree/compiler/setup.py.in
@@ -31,7 +31,7 @@ replace the `pyiree.compiler` and `pyiree.tf.compiler` packages.
 ## Core compiler
 
 ```py
-from pyiree.compiler2 import *
+from pyiree.compiler import *
 
 SIMPLE_MUL_ASM = """
 func @simple_mul(%arg0: tensor<4xf32>, %arg1: tensor<4xf32>) -> tensor<4xf32>
@@ -43,7 +43,7 @@ func @simple_mul(%arg0: tensor<4xf32>, %arg1: tensor<4xf32>) -> tensor<4xf32>
 
 # Also see compile_file()
 # There are many keyword options available.
-# See pyiree.compiler2.CompilerOptions
+# See pyiree.compiler.CompilerOptions
 binary = compile_str(SIMPLE_MUL_ASM, target_backends=["vulkan-spirv"])
 ```
 
@@ -52,7 +52,7 @@ binary = compile_str(SIMPLE_MUL_ASM, target_backends=["vulkan-spirv"])
 
 ```py
 import tensorflow as tf
-from pyiree.compiler2.tf import *
+from pyiree.compiler.tf import *
 
 class SimpleArithmeticModule(tf.Module):
 
@@ -65,7 +65,7 @@ class SimpleArithmeticModule(tf.Module):
 
 # Also see compile_saved_model to directly compile an on-disk saved model.
 # There are many keyword options available.
-# See: pyiree.compiler2.tf.ImportOptions
+# See: pyiree.compiler.tf.ImportOptions
 binary = compile_module(
     SimpleArithmeticModule(), target_backends=["vulkan-spirv"])
 ```
@@ -125,14 +125,13 @@ setup(
     ],
     python_requires=">=3.6",
     packages=find_namespace_packages(
-        include=["pyiree.compiler2", "pyiree.compiler2.*", "pyiree.tools.core"
-                ]),
+        include=["pyiree.compiler", "pyiree.compiler.*", "pyiree.tools.core"]),
     package_data={
         "pyiree.tools.core": [f"iree-translate{exe_suffix}",],
     },
     cmdclass={
-      'bdist_wheel': bdist_wheel,
-      'install': platlib_install,
+        'bdist_wheel': bdist_wheel,
+        'install': platlib_install,
     },
     zip_safe=False,  # This package is fine but not zipping is more versatile.
 )
