+++ bindings/python/pyiree/rt/vm_test.py
@@ -18,7 +18,7 @@
 from absl import logging
 from absl.testing import absltest
 import numpy as np
-from pyiree import compiler2 as compiler
+from pyiree import compiler
 from pyiree import rt
