+++ bindings/python/tests/compiler_tf_test.py
@@ -20,7 +20,7 @@ import unittest
 
 # TODO: No idea why pytype cannot find names from this module.
 # pytype: disable=name-error
-from pyiree.compiler2.tf import *
+from pyiree.compiler.tf import *
 
 if not is_available():
   print(f"Skipping test {__file__} because the IREE TensorFlow compiler "
