+++ bindings/python/tests/compiler_tflite_test.py
@@ -21,7 +21,7 @@ import unittest
 
 # TODO: No idea why pytype cannot find names from this module.
 # pytype: disable=name-error
-from pyiree.compiler2.tflite import *
+from pyiree.compiler.tflite import *
 
 if not is_available():
   print(f"Skipping test {__file__} because the IREE TFLite compiler "
